Rich text: move format boundaries to ref callback and separate file
